axi_laser_driver: Add control logic for the TIA channel selection lines
Implement a sequencer which will generate predefined values for the TIA
channel selection lines.

The sequencer has an auto and a manual mode, and can be set to custom
sequences of the TIA channel selection lines sate.

The sequencer in auto mode is synchronized to the pulse, it will change
its state before a generated pulse which will drive the lasers. The
offset between the sequencer beat and the laser driver pulse can be
modified through an AXI register.

The following new registers were introduced:

   0xAC - SEQUENCER_CONTROL
          0 - SEQUENCER_ENABLE
          1 - AUTO_SEQUENCER_ENABLED
   0xB0 - SEQUENCER_SYNC_OFFSET
   0xB4 - AUTO_SEQUENCE
          [ 1: 0] - CHANNEL_SEL_0
          [ 5: 4] - CHANNEL_SEL_1
          [ 9: 8] - CHANNEL_SEL_2
          [13:12] - CHANNEL_SEL_3
   0xB8 - MANUAL_SEQUENCE
          [ 1: 0] - MANUAL_CHANNEL_SEL